rename: 패키지 및 메인 클래스 이름 변경
- 패키지 이름에 사용된 도메인은 임시로 wink 공식 홈페이지로 작성
