+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'back-end'
+rootProject.name = 'PlanList-backend'
