+++ src/main/resources/application.properties
@@ -1,1 +1,1 @@
-spring.application.name=back-end
+spring.application.name=PlanList-backend
