+++ src/test/java/kr/ac/kookmin/wink/planlist/PlanListApplicationTests.java
@@ -1,10 +1,10 @@
-package com.backend;
+package kr.ac.kookmin.wink.planlist;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BackEndApplicationTests {
+class PlanListApplicationTests {
 
     @Test
     void contextLoads() {
